requirements.txt
pyrebase4
google-cloud-firestore
gcloud
setuptools